% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do_PCoA.R
\name{do_PCoA}
\alias{do_PCoA}
\title{Perform PCoA on a distance matrix}
\usage{
do_PCoA(DistanceMatrix)
}
\arguments{
\item{DistanceMatrix}{a dist object containing the distances between samples}
}
\description{
This is a convenience wrapper for the ‘cmdscale’ function from the vegan
package. It takes a distance matrix (an object of type "dist") and performs a
PCoA ordination with ‘cmdscale’, for two dimensions (k = 2). Returns a list
containing ‘Coordinates’, a tbl with the PCoA coordinates; and ‘Variance1’
and ‘Variance2’, the percentage variance explained by the first and second
PCoA axes respectively.

The method for calculating variance explained comes from
\url{http://r-sig-ecology.471788.n2.nabble.com/Variability-explanations-for-the-PCO-axes-as-in-Anderson-and-Willis-2003-td6429547.html}.
}
\seealso{
dist for info on the dist object

read_dist to read a distance matrix from a file (e.g. reading in
UniFrac distances)

cmdscale the vegan function that performs the actual ordination
}

